pip-requirements.txt
Flask==0.10.1
Flask-Gravatar==0.4.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.3
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
SQLAlchemy==0.9.8
WTForms==2.0.1
Werkzeug==0.9.6
amqp==1.4.6
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.19
cached-property==0.1.5
celery==3.1.17
enum==0.4.4
itsdangerous==0.24
kombu==3.0.24
psycopg2==2.5.4
python-ldap==2.4.18
python-memcached==1.53
pytz==2014.10
requests==2.5.1
sqltap==0.3.5
wsgiref==0.1.2
